Fix counting of lines: \n is need, not \\n.
Unlike \w, \n is not a regexp special in elisp. Hence, with "\\n", the
regexp machinery strips the backslash (as it does with all unrecognized
specials), and we're actually counting the character 'n'. If we use "\n"
instead, that's a real newline character there, which is the right thing to
count.