Allow wc stat generation to be deferred to idle periods
This commit adds `defcustom wc-idle-wait`, and sets it to 0 for
backwards compatibility.  It also adds three local variables.  The first
is the buffer-local `wc-timer-tracker`, which sets up one timer for
each buffer wc-mode is enabled for.  In the future stat generation
will be suspended in inactive buffers.

`run-with-idle-timer` does not appear to function correctly without
lexical-binding; it also does not function correctly when the FUNCTION
in its form (run-with-idle-timer SECS REPEAT FUNCTION &rest ARGS) is
used to call wc-mode-update.  The limitations of `run-with-idle-timer`
appears to be related to an inability to handle any side-effects of
FUNCTION.  This limitation is worked around using the
`wc-buffer-stats` variable, which is where the mode-line now draws its
data from.

Finally, the provided kill-buffer-hook cleans up each buffer-local
`wc-timer-tracker` variable at the appropriate time.

This modified design should also facilitate enabling asynchronous stat
collection at a later date.